Extract anti_pattern_flow.rs from turn.rs (parent #680)
Vertical-slice extraction: move the anti-pattern extraction + retrieval
flow out of turn.rs into a new src/agent/loop_run/anti_pattern_flow.rs
sibling module. First conversion of `impl Agent` methods to free
functions in this PR series; matches the existing actor_loop_flow /
verifier_orchestration pattern.

Migrated (all pub(super) free fns over &mut Agent / &Agent):

Entry points:
- maybe_extract_anti_pattern (Issue #464 post-loop upsert)
- try_inject_anti_pattern_message (Issue #464 pre-prompt retrieval)

Private helpers:
- anti_pattern_extraction_feedback (eligibility gate)
- finish_anti_pattern_extraction (per-turn cap setter)
- log_anti_pattern_skip / log_anti_pattern_extracted /
  log_anti_pattern_extract_result (telemetry)

No facade re-export (DR3-001). Caller updates:
- turn.rs: `self.try_inject_anti_pattern_message()` →
  `super::anti_pattern_flow::try_inject_anti_pattern_message(self)`
- actor_loop_flow.rs: `agent.maybe_extract_anti_pattern()` →
  `super::anti_pattern_flow::maybe_extract_anti_pattern(agent)`

`cargo fix --tests` pruned 2 newly-unused imports from turn.rs's
prelude (`FeedbackFrame`, the `FeedbackKind` variant import line).

turn.rs: 7,172 -> 6,885 (-287 LOC)
Cumulative from 39,489: -32,604 LOC (-82.6%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
