+++ src/agent/loop_run/turn.rs
@@ -64,8 +64,7 @@ use super::safe_stop_payload::{build_safe_stop_payload, collect_recent_action_la
 #[cfg(debug_assertions)]
 use super::small_helpers::masked_path_hash_bounded_list;
 use super::small_helpers::{
-    anti_pattern_failed_action_summary, latest_tool_result_since_last_user, raw_mode_safe_text,
-    rfc3339_now_utc, user_interrupt_result,
+    latest_tool_result_since_last_user, raw_mode_safe_text, rfc3339_now_utc, user_interrupt_result,
 };
 use super::spinner::{Spinner, SpinnerStopSignal};
 use super::summary::{ExitReason, LoopResult};
@@ -124,7 +123,7 @@ use crate::logging::{log_llm_event, stable_path_hash};
 use crate::model_capabilities::model_capabilities;
 use crate::modes::plan_act::{ModeClassification, WorkMode, classify_work_mode_json};
 use crate::ollama::xml_fallback::normalize_tool_call_arguments;
-use crate::session::feedback::{FeedbackFrame, FeedbackKind};
+use crate::session::feedback::FeedbackKind;
 use crate::session::precaution::PrecautionStatus;
 use crate::session::store::ScaffoldArtifactFileSnapshot;
 use crate::tools::registry::{BashErrorClass, ToolSpec};
@@ -1196,293 +1195,6 @@ impl Agent {
             }
         }
     }
-
-    /// Issue #464: post-loop AntiPattern extraction. Mirrors
-    /// `maybe_extract_case_record` but triggers on **failure** turns instead
-    /// of success. Upserts a record keyed by (workspace_key, task_signature,
-    /// feedback_kind); the second + N-th occurrence increments `repeat_count`.
-    /// Pure upsert / scrub / persist; no sidecar / LLM calls.
-    pub(super) fn maybe_extract_anti_pattern(&mut self) {
-        use crate::session::anti_pattern;
-
-        let Some(frame) = self.anti_pattern_extraction_feedback() else {
-            return;
-        };
-        let kind = frame.kind.clone();
-
-        if anti_pattern::anti_pattern_dry_run(|k| std::env::var(k)) {
-            self.log_anti_pattern_skip("dry_run", Some(&kind), None);
-            self.finish_anti_pattern_extraction();
-            return;
-        }
-
-        let summary_owned = anti_pattern_failed_action_summary(&frame);
-        let language_stack = derive_language_stack(&self.work_root);
-        let active_task = self.session.working_memory.active_task.clone();
-        let workspace_key = self.session.workspace_key.clone();
-        let touched_files = self.session.working_memory.touched_files.clone();
-        let inputs = anti_pattern::AntiPatternRecordInputs {
-            workspace_key: &workspace_key,
-            work_root: &self.work_root,
-            active_task: active_task.as_deref(),
-            language_stack: &language_stack,
-            touched_files: &touched_files,
-            feedback_kind: kind.clone(),
-            failed_action_summary: &summary_owned,
-        };
-
-        let started = std::time::Instant::now();
-        let state_root = self.session_store.state_root().to_path_buf();
-        self.log_anti_pattern_extract_result(
-            anti_pattern::extract_or_increment(&state_root, &inputs),
-            started,
-        );
-        self.finish_anti_pattern_extraction();
-    }
-
-    fn anti_pattern_extraction_feedback(&mut self) -> Option<FeedbackFrame> {
-        use crate::session::anti_pattern;
-
-        if self.session.mode_state.mode == ExecutionMode::Plan {
-            return None;
-        }
-        if self.session.anti_pattern_extracted_this_turn {
-            return None;
-        }
-        if anti_pattern::anti_pattern_disabled(|k| std::env::var(k)) {
-            log_llm_event(
-                "agent.anti_pattern.disabled",
-                serde_json::json!({
-                    "session_id": self.session_store.session_id(),
-                }),
-            );
-            self.finish_anti_pattern_extraction();
-            return None;
-        }
-        let Some(frame) = self.session.last_feedback.clone() else {
-            self.finish_anti_pattern_extraction();
-            return None;
-        };
-        if !anti_pattern::is_repeat_eligible_kind(&frame.kind) {
-            self.finish_anti_pattern_extraction();
-            return None;
-        }
-        Some(frame)
-    }
-
-    fn finish_anti_pattern_extraction(&mut self) {
-        self.session.anti_pattern_extracted_this_turn = true;
-    }
-
-    fn log_anti_pattern_skip(
-        &self,
-        reason: &str,
-        feedback_kind: Option<&FeedbackKind>,
-        bytes: Option<usize>,
-    ) {
-        let mut payload = serde_json::json!({
-            "session_id": self.session_store.session_id(),
-            "reason": reason,
-        });
-        if let Some(kind) = feedback_kind {
-            payload["feedback_kind"] = serde_json::to_value(kind).unwrap_or_default();
-        }
-        if let Some(bytes) = bytes {
-            payload["bytes"] = serde_json::json!(bytes);
-        }
-        log_llm_event("agent.anti_pattern.skipped", payload);
-    }
-
-    fn log_anti_pattern_extracted(
-        &self,
-        record: &crate::session::anti_pattern::AntiPatternRecord,
-        outcome: &str,
-        started: std::time::Instant,
-    ) {
-        let compute_ms = started.elapsed().as_secs_f64() * 1000.0;
-        log_llm_event(
-            "agent.anti_pattern.extracted",
-            serde_json::json!({
-                "session_id": self.session_store.session_id(),
-                "anti_pattern_id": record.anti_pattern_id,
-                "outcome": outcome,
-                "repeat_count": record.repeat_count,
-                "feedback_kind": serde_json::to_value(&record.feedback_kind).unwrap_or_default(),
-                "compute_ms": compute_ms,
-            }),
-        );
-    }
-
-    fn log_anti_pattern_extract_result(
-        &self,
-        result: Result<
-            crate::session::anti_pattern::ExtractOutcome,
-            crate::session::anti_pattern::PersistError,
-        >,
-        started: std::time::Instant,
-    ) {
-        use crate::session::anti_pattern::{ExtractOutcome, PersistError};
-
-        match result {
-            Ok(ExtractOutcome::Created(record)) => {
-                self.log_anti_pattern_extracted(&record, "created", started);
-            }
-            Ok(ExtractOutcome::Incremented(record)) => {
-                self.log_anti_pattern_extracted(&record, "incremented", started);
-            }
-            Ok(ExtractOutcome::SkippedIneligibleKind) => {
-                self.log_anti_pattern_skip("ineligible_kind", None, None);
-            }
-            Ok(ExtractOutcome::SkippedNoActiveTask) => {
-                self.log_anti_pattern_skip("no_active_task", None, None);
-            }
-            Err(PersistError::TooLarge { bytes }) => {
-                self.log_anti_pattern_skip("too_large", None, Some(bytes));
-            }
-            Err(err) => {
-                log_llm_event(
-                    "agent.anti_pattern.failed",
-                    serde_json::json!({
-                        "session_id": self.session_store.session_id(),
-                        "error": err.to_string(),
-                    }),
-                );
-            }
-        }
-    }
-
-    /// Issue #464: build and (when applicable) inject an `Avoid Patterns:`
-    /// system message into the next prompt. Mirrors
-    /// `try_inject_case_retrieval_message` but pulls from
-    /// `state_root/anti_patterns/`.
-    pub(super) fn try_inject_anti_pattern_message(&mut self) -> Option<RetrievalInjection> {
-        use crate::session::anti_pattern::{self, AntiPatternRetrievalInputs, RetrievalOutcome};
-        use crate::session::case_record::{build_task_signature, capture_repo_fingerprint};
-
-        // 1. Plan mode → skipped, do not consume cap.
-        if self.session.mode_state.mode == ExecutionMode::Plan {
-            log_llm_event(
-                "agent.anti_pattern.skipped",
-                serde_json::json!({
-                    "session_id": self.session_store.session_id(),
-                    "reason": "plan_mode",
-                }),
-            );
-            return None;
-        }
-        // 2. per-turn cap consumed.
-        if self.session.anti_pattern_retrieval_invoked_this_turn {
-            log_llm_event(
-                "agent.anti_pattern.skipped",
-                serde_json::json!({
-                    "session_id": self.session_store.session_id(),
-                    "reason": "per_turn_cap_consumed",
-                }),
-            );
-            return None;
-        }
-        // 3. Env disable.
-        if anti_pattern::anti_pattern_disabled(|k| std::env::var(k)) {
-            self.session.anti_pattern_retrieval_invoked_this_turn = true;
-            log_llm_event(
-                "agent.anti_pattern.disabled",
-                serde_json::json!({
-                    "session_id": self.session_store.session_id(),
-                }),
-            );
-            return None;
-        }
-
-        // 4. Build inputs.
-        let language_stack = derive_language_stack(&self.work_root);
-        let workspace_key = self.session.workspace_key.clone();
-        let active_task = self.session.working_memory.active_task.clone();
-        let task_signature = build_task_signature(active_task.as_deref(), &self.work_root);
-        let repo_fp = capture_repo_fingerprint(&workspace_key, &self.work_root, &language_stack);
-        let touched_files = self.session.working_memory.touched_files.clone();
-        let feedback_kind = self.session.last_feedback.as_ref().map(|f| f.kind.clone());
-
-        let dry_run = anti_pattern::anti_pattern_dry_run(|k| std::env::var(k));
-        let inputs = AntiPatternRetrievalInputs {
-            current_task_signature: &task_signature,
-            current_language_stack: &language_stack,
-            current_repo_fingerprint: &repo_fp,
-            current_touched_files: &touched_files,
-            current_feedback_kind: feedback_kind,
-        };
-
-        self.session.anti_pattern_retrieval_invoked_this_turn = true;
-
-        let state_root = self.session_store.state_root().to_path_buf();
-        match anti_pattern::retrieve_relevant_anti_patterns(&state_root, &inputs, dry_run) {
-            Ok(RetrievalOutcome::Completed {
-                candidate_count,
-                selected,
-                skipped_corrupt_count,
-                compute_ms,
-            }) => {
-                let top_score = selected.first().map(|s| s.breakdown.total).unwrap_or(0.0);
-                let top_id = selected
-                    .first()
-                    .map(|s| s.record.anti_pattern_id.clone())
-                    .unwrap_or_default();
-                log_llm_event(
-                    "agent.anti_pattern.retrieved",
-                    serde_json::json!({
-                        "session_id": self.session_store.session_id(),
-                        "candidate_count": candidate_count,
-                        "selected_count": selected.len(),
-                        "top_score": top_score,
-                        "top_anti_pattern_id": top_id,
-                        "threshold": anti_pattern::ANTI_PATTERN_RETRIEVAL_SCORE_THRESHOLD,
-                        "compute_ms": compute_ms,
-                        "skipped_corrupt_count": skipped_corrupt_count,
-                    }),
-                );
-                // Issue #555: capture selected IDs for photon mapper before
-                // format_for_prompt consumes `selected`.
-                let selected_ids: Vec<String> = selected
-                    .iter()
-                    .map(|s| s.record.anti_pattern_id.clone())
-                    .collect();
-                anti_pattern::format_for_prompt(&selected)
-                    .map(ConversationMessage::system)
-                    .map(|message| RetrievalInjection {
-                        message,
-                        selected_ids,
-                    })
-            }
-            Ok(RetrievalOutcome::Skipped {
-                reason,
-                candidate_count,
-                skipped_corrupt_count,
-                compute_ms,
-            }) => {
-                log_llm_event(
-                    "agent.anti_pattern.skipped",
-                    serde_json::json!({
-                        "session_id": self.session_store.session_id(),
-                        "reason": reason.as_log_str(),
-                        "candidate_count": candidate_count,
-                        "skipped_corrupt_count": skipped_corrupt_count,
-                        "compute_ms": compute_ms,
-                    }),
-                );
-                None
-            }
-            Err(error) => {
-                log_llm_event(
-                    "agent.anti_pattern.failed",
-                    serde_json::json!({
-                        "session_id": self.session_store.session_id(),
-                        "error": error,
-                    }),
-                );
-                None
-            }
-        }
-    }
-
     fn run_turn(
         &mut self,
         input: &str,
@@ -3086,7 +2798,7 @@ impl Agent {
         if let Some(ref inj) = case_injection {
             messages.push(inj.message.clone());
         }
-        let anti_injection = self.try_inject_anti_pattern_message();
+        let anti_injection = super::anti_pattern_flow::try_inject_anti_pattern_message(self);
         if let Some(ref inj) = anti_injection {
             messages.push(inj.message.clone());
         }
